Tuesday 25th November - Project Setup
Date: Thursday 13 November
Lesson number:1
1. Goal for today (C i)
 State what your plan said you should complete during this session. My goal for this lesson was to complete the setup of the system and store it in the registration template.

2. Work completed (C iii)
 Write one or two sentences summarising what you achieved in this session.
Since I spent most of the lesson completing my flowchart assessment, I did not have a lot of time to finish the setup, so I still need to finish the basic page structure.
3. Did you follow your plan? (C iii and C iv)
No: I started my plan, but did not have the time to finish it.
4. Next step (C i)
 List one or two tasks you intend to complete during the next lesson.
For the next lesson, I plan to finish completing the basic setup for the registration system and write the required fields.

Date: Tuesday 25th November
Lesson number: 4
1. Goal for today (C i)
 State what your plan said you should complete during this session.
 My goal for this lesson was to complete the different fields on the registration form.

2. Work completed (C iii)
 Write one or two sentences summarising what you achieved in this session.
I did  achieve my goal, since I completed all the required fields that my registration form needs to have.
3. Did you follow your plan? (C iii and C iv)
Yes, I followed through with my plan.
4. Next step (C i)
 List one or two tasks you intend to complete during the next lesson.
For the next lesson, I plan to finish completing the dropdown boxes that some of my fields (e.g. date of birth and nationality) need to have, as well as inputting icons on the webpage this way it is easier to navigate.